Gather enum definitions
