+++ lax/src/svddc.rs
@@ -2,26 +2,6 @@ use crate::{error::*, layout::MatrixLayout, *};
 use cauchy::*;
 use num_traits::{ToPrimitive, Zero};
 
-/// Specifies how many of the columns of *U* and rows of *V*ᵀ are computed and returned.
-///
-/// For an input array of shape *m*×*n*, the following are computed:
-#[derive(Clone, Copy, Eq, PartialEq)]
-#[repr(u8)]
-pub enum UVTFlag {
-    /// All *m* columns of *U* and all *n* rows of *V*ᵀ.
-    Full = b'A',
-    /// The first min(*m*,*n*) columns of *U* and the first min(*m*,*n*) rows of *V*ᵀ.
-    Some = b'S',
-    /// No columns of *U* or rows of *V*ᵀ.
-    None = b'N',
-}
-
-impl UVTFlag {
-    fn as_ptr(&self) -> *const i8 {
-        self as *const UVTFlag as *const i8
-    }
-}
-
 pub trait SVDDC_: Scalar {
     fn svddc(l: MatrixLayout, jobz: UVTFlag, a: &mut [Self]) -> Result<SVDOutput<Self>>;
 }
